pylint updates
